+++ graphene/types/tests/test_mutation.py
@@ -8,9 +8,6 @@ from ..scalars import String
 from ..schema import Schema
 from ..structures import NonNull
 from ..interface import Interface
-from ..enum import Enum
-from ..inputobjecttype import InputObjectType
-from graphene.compat.middleware import enum_value_convertor_middleware
 
 
 class MyType(Interface):
@@ -221,188 +218,3 @@ def test_mutation_as_subclass():
     )
     assert not result.errors
     assert result.data == {"createUserWithPlanet": {"name": "Peter", "planet": "earth"}}
-
-
-def test_mutation_enum_input():
-    class RGB(Enum):
-        """Available colors"""
-
-        RED = 1
-        GREEN = 2
-        BLUE = 3
-
-    color_input = None
-
-    class CreatePaint(Mutation):
-        class Arguments:
-            color = RGB(required=True)
-
-        color = RGB(required=True)
-
-        def mutate(_, info, color):
-            nonlocal color_input
-            color_input = color
-            return CreatePaint(color=color)
-
-    class MyMutation(ObjectType):
-        create_paint = CreatePaint.Field()
-
-    class Query(ObjectType):
-        a = String()
-
-    schema = Schema(query=Query, mutation=MyMutation)
-    result = schema.execute(
-        """ mutation MyMutation {
-        createPaint(color: RED) {
-            color
-        }
-    }
-    """
-    )
-    assert not result.errors
-    assert result.data == {"createPaint": {"color": "RED"}}
-
-    assert color_input == RGB.RED
-
-
-def test_mutation_enum_input_type():
-    class RGB(Enum):
-        """Available colors"""
-
-        RED = 1
-        GREEN = 2
-        BLUE = 3
-
-    class ColorInput(InputObjectType):
-        color = RGB(required=True)
-
-    color_input_value = None
-
-    class CreatePaint(Mutation):
-        class Arguments:
-            color_input = ColorInput(required=True)
-
-        color = RGB(required=True)
-
-        def mutate(_, info, color_input):
-            nonlocal color_input_value
-            color_input_value = color_input.color
-            return CreatePaint(color=color_input.color)
-
-    class MyMutation(ObjectType):
-        create_paint = CreatePaint.Field()
-
-    class Query(ObjectType):
-        a = String()
-
-    schema = Schema(query=Query, mutation=MyMutation)
-    result = schema.execute(
-        """ mutation MyMutation {
-        createPaint(colorInput: { color: RED }) {
-            color
-        }
-    }
-    """,
-    )
-    assert not result.errors
-    assert result.data == {"createPaint": {"color": "RED"}}
-
-    assert color_input_value == RGB.RED
-
-
-def test_mutation_enum_input_compatability_middleware():
-    """Test the `enum_value_convertor_middleware`"""
-
-    class RGB(Enum):
-        """Available colors"""
-
-        RED = 1
-        GREEN = 2
-        BLUE = 3
-
-    color_input = None
-
-    class CreatePaint(Mutation):
-        class Arguments:
-            color = RGB(required=True)
-
-        color = RGB(required=True)
-
-        def mutate(_, info, color):
-            nonlocal color_input
-            color_input = color
-            return CreatePaint(color=color)
-
-    class MyMutation(ObjectType):
-        create_paint = CreatePaint.Field()
-
-    class Query(ObjectType):
-        a = String()
-
-    schema = Schema(query=Query, mutation=MyMutation)
-    result = schema.execute(
-        """ mutation MyMutation {
-        createPaint(color: RED) {
-            color
-        }
-    }
-    """,
-        middleware=[enum_value_convertor_middleware],
-    )
-    assert not result.errors
-    assert result.data == {"createPaint": {"color": "RED"}}
-
-    assert color_input == 1
-    assert type(color_input) == int
-
-
-def test_mutation_enum_input_compatability_middleware_input_type():
-    """Test the `enum_value_convertor_middleware`"""
-
-    class RGB(Enum):
-        """Available colors"""
-
-        RED = 1
-        GREEN = 2
-        BLUE = 3
-
-    class SecondColorInput(InputObjectType):
-        color = RGB(required=True)
-
-    class ColorInput(InputObjectType):
-        color_input = SecondColorInput(required=True)
-
-    color_input_value = None
-
-    class CreatePaint(Mutation):
-        class Arguments:
-            color_input = ColorInput(required=True)
-
-        color = RGB(required=True)
-
-        def mutate(_, info, color_input):
-            nonlocal color_input_value
-            color_input_value = color_input.color_input.color
-            return CreatePaint(color=color_input_value)
-
-    class MyMutation(ObjectType):
-        create_paint = CreatePaint.Field()
-
-    class Query(ObjectType):
-        a = String()
-
-    schema = Schema(query=Query, mutation=MyMutation)
-    result = schema.execute(
-        """ mutation MyMutation {
-        createPaint(colorInput: { colorInput: { color: RED } }) {
-            color
-        }
-    }
-    """,
-        middleware=[enum_value_convertor_middleware],
-    )
-    assert not result.errors
-    assert result.data == {"createPaint": {"color": "RED"}}
-
-    assert color_input_value == 1
-    assert type(color_input_value) == int
